Refactor help menu items to status-bar module
Moved help menu constants and HelpMenuItemsLabel component from the project OverviewCard to the status-bar module. This improves code organization by colocating help menu logic with the status bar, where it is used.
